Struct termspin::Group

pub struct Group { /* private fields */ }

A stateful group of displayable frames that are separated by new lines.

Implementations

impl Group

pub fn new() -> Self

Create a new empty group.

pub fn len(&self) -> usize

The amount of items in this group.

pub fn is_empty(&self) -> bool

Whether the group has no children.

pub fn insert(&mut self, idx: usize, frames: impl Frames) -> &mut Self

Insert an item at the given position.

Panics

Panics if index > len.

pub fn push(&mut self, frames: impl Frames) -> &mut Self

Push an item at the end.

pub fn extend<I>(&mut self, iter: I) -> &mut Selfwhere I: IntoIterator, I::Item: Frames,

Extend this group from an iterator.

pub fn remove(&mut self, idx: usize) -> &mut Self

Remove an item at the given position. No-op if the position is invalid.

pub fn iter(&self) -> impl Iterator<Item = &dyn Frames> + '_

Return an iterator of the frames in this group.

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut dyn Frames> + '_

Return an iterator of the frames in this group.

pub fn retain(&mut self, f: impl Fn(&dyn Frames) -> bool)

Retains only the items specified by the predicate.

pub fn clear(&self) -> impl Display + '_

Print ANSI codes that clears the frames displayed by this group.

Example

print!("{}", group.clear());

pub fn get_indent(&self) -> usize

Return the indentation level of this group.

pub fn with_indent(self, level: usize) -> Self

Set the indentation level of this group.

pub fn set_indent(&mut self, level: usize) -> &mut Self

Set the indentation level of this group.

pub fn shared(self) -> SharedFrames<Self>

Turn this group into SharedFrames.