Trait termspin::Frames

pub trait Frames: AnySync + Display {
    // Required method
    fn advance(&mut self);

    // Provided methods
    fn reset(&mut self) { ... }
    fn clear(&self, _f: &mut Formatter<'_>) -> Result { ... }
    fn lines(&self) -> usize { ... }
    fn print_len(&self) -> Option<usize> { ... }
}

Frames that can be printed to the terminal via fmt::Display.

The printed text should not end with a new line.

Required Methods

fn advance(&mut self)

Advance to the next frame.

Provided Methods

fn reset(&mut self)

Reset to the first frame.

fn clear(&self, _f: &mut Formatter<'_>) -> Result

Write ANSI codes to the given formatter that clears the printed output.

fn lines(&self) -> usize

The amount of lines that is supposed to be printed and cleared.

This is required by groups.

fn print_len(&self) -> Option<usize>

The length of the printed text if known in advance.

Implementations

impl dyn Frames

pub fn is<_T>(&self) -> boolwhere _T: Any, Self: Downcast<_T>,

pub fn downcast_ref<_T>(&self) -> Result<&_T, TypeMismatch>where _T: Any, Self: Downcast<_T>,

pub fn downcast_mut<_T>(&mut self) -> Result<&mut _T, TypeMismatch>where _T: Any, Self: Downcast<_T>,

pub fn downcast<_T>( self: Box<Self> ) -> Result<Box<_T>, DowncastError<Box<Self>>>where _T: Any, Self: Downcast<_T>,

pub fn downcast_rc<_T>( self: Rc<Self> ) -> Result<Rc<_T>, DowncastError<Rc<Self>>>where _T: Any, Self: Downcast<_T>,

pub fn downcast_arc<_T>( self: Arc<Self> ) -> Result<Arc<_T>, DowncastError<Arc<Self>>>where _T: AnySync, Self: DowncastSync<_T>,