Struct termspin::Loop

pub struct Loop<F: Frames> { /* private fields */ }

Run the loop with the given callback.

Example

let l = Loop::new(
    Duration::from_millis(100),
    Line::new(spinner::from_iter([r"\", "|", "/"]))
);

// Run the loop while blocking the current thread.
l.run(|out| print!("{out}"));

Implementations

impl<F: Frames> Loop<F>

pub fn new(interval: Duration, frames: F) -> Self

Create a new loop that updates at the given interval.

pub fn run(&self, f: impl FnMut(&dyn Display) -> Result<()>) -> Result<()>

Run the loop with the given callback.

Example

let l = Loop::new(
    Duration::from_millis(100),
    Line::new(spinner::from_iter([r"\", "|", "/"]))
);

// Run the loop while blocking the current thread.
l.run(|out| print!("{out}"));

pub fn run_stream(&self, stream: impl Write) -> Result<()>

Run the loop outputting frames to the given stream.

pub fn clear_stream(&self, stream: impl Write) -> Result<()>

A convenience function to clear the given stream.

pub fn spawn_stream<S>(&self, stream: S)where S: Write + Send + 'static,

Spawn the loop on a separate thread, no-op if the loop is already running.

pub fn stop(&self)

Stop a running loop.

pub fn wait(&self, duration: Duration)

Wait for the given duration before the next cycle.

pub fn reset(&self)

Reset the frames of the loop.

pub fn inner(&self) -> Fwhere F: Clone,

Clone the inner frames.

pub fn auto_stop(&self, stop: bool)

Exit the running loop if only one instance of the loop exists, defaults to true.

It is useful when you spawn the spawn the loop on a separate thread that should exit when all handles to it go out of scope.